Faster System Tools menu loading
The database will not be re-created at each launch.
Run "System Tools -> USER INTERFACE -> Refresh System Tools Menu ".

+++ Apps/SystemTools/launch.sh
@@ -4,15 +4,6 @@ echo $0 $*
 echo performance >/sys/devices/system/cpu/cpu0/cpufreq/scaling_governor
 echo 1416000 >/sys/devices/system/cpu/cpu0/cpufreq/scaling_min_freq
 
-/mnt/SDCARD/System/bin/sdl2imgshow \
-  -i "/mnt/SDCARD/trimui/res/crossmix-os/bg-info.png" \
-  -f "/mnt/SDCARD/System/resources/DejaVuSans.ttf" \
-  -s 50 \
-  -c "220,220,220" \
-  -t "Building Menu..." &
-sleep 0.2
-pkill -f sdl2imgshow
-
 PATH="/mnt/SDCARD/System/bin:$PATH"
 export LD_LIBRARY_PATH="/mnt/SDCARD/System/lib:/usr/trimui/lib:$LD_LIBRARY_PATH"
 database_file="/mnt/SDCARD/Apps/SystemTools/Menu/Menu_cache7.db"
@@ -23,10 +14,36 @@ if [ ! -f "/mnt/SDCARD/System/etc/systemtools.json" ]; then
   touch /mnt/SDCARD/System/etc/systemtools.json
 fi
 
-SubFoldersList=":"
 ####################################### For testing :
-rm "$database_file"
+# rm "$database_file"
 #######################################
+
+rebuildmenu=false
+for arg in "$@"; do
+  if [ "$arg" = "-rebuildmenu" ]; then
+    rebuildmenu=true
+    break
+  fi
+done
+
+if [ "$rebuildmenu" = true ]; then
+  rm "$database_file"
+  sync
+  LaunchMessage="Building Menu..."
+else
+  LaunchMessage="Loading..."
+fi
+
+/mnt/SDCARD/System/bin/sdl2imgshow \
+  -i "/mnt/SDCARD/trimui/res/crossmix-os/bg-info.png" \
+  -f "/mnt/SDCARD/System/resources/DejaVuSans.ttf" \
+  -s 50 \
+  -c "220,220,220" \
+  -t "$LaunchMessage" &
+sleep 0.2
+pkill -f sdl2imgshow
+
+SubFoldersList=":"
 
 cp /mnt/SDCARD/Apps/SystemTools/GoTo_SystemTools_List.json /tmp/state.json
 sync
